refactor(v2): remove v2 compatible related code (#7103)
* refactor(v2): remove v2 compatible related code

* fix presubmit

* clean up more usages of v2 compatible

* fix
