+++ v2/component/launcher_v2.go
@@ -4,12 +4,17 @@ import (
 	"bytes"
 	"context"
 	"encoding/json"
+	"errors"
 	"fmt"
 	"io/ioutil"
+	"os"
+	"os/exec"
+	"path/filepath"
 	"strconv"
 	"strings"
 	"time"
 
+	"github.com/golang/glog"
 	"github.com/golang/protobuf/ptypes/timestamp"
 	"github.com/kubeflow/pipelines/v2/cacheutils"
 	api "github.com/kubeflow/pipelines/v2/kfp-api"
@@ -289,4 +294,371 @@ func prettyPrint(jsonStr string) string {
 		return jsonStr
 	}
 	return string(prettyJSON.Bytes())
+}
+
+const OutputMetadataFilepath = "/tmp/kfp_outputs/output_metadata.json"
+
+func execute(ctx context.Context, executorInput *pipelinespec.ExecutorInput, cmd string, args []string, bucket *blob.Bucket, bucketConfig *objectstore.Config, namespace string, k8sClient *kubernetes.Clientset) (*pipelinespec.ExecutorOutput, error) {
+	if err := downloadArtifacts(ctx, executorInput, bucket, bucketConfig, namespace, k8sClient); err != nil {
+		return nil, err
+	}
+	if err := prepareOutputFolders(executorInput); err != nil {
+		return nil, err
+	}
+
+	// Fill in placeholders with runtime values.
+	placeholders, err := getPlaceholders(executorInput)
+	if err != nil {
+		return nil, err
+	}
+	for placeholder, replacement := range placeholders {
+		cmd = strings.ReplaceAll(cmd, placeholder, replacement)
+	}
+	for i := range args {
+		arg := args[i]
+		for placeholder, replacement := range placeholders {
+			arg = strings.ReplaceAll(arg, placeholder, replacement)
+		}
+		args[i] = arg
+	}
+
+	// Run user program.
+	executor := exec.Command(cmd, args...)
+	executor.Stdin = os.Stdin
+	executor.Stdout = os.Stdout
+	executor.Stderr = os.Stderr
+	defer glog.Flush()
+	if err := executor.Run(); err != nil {
+		return nil, err
+	}
+
+	// Collect outputs from output metadata file.
+	return getExecutorOutputFile(executorInput.GetOutputs().GetOutputFile())
+}
+
+type uploadOutputArtifactsOptions struct {
+	bucketConfig   *objectstore.Config
+	bucket         *blob.Bucket
+	metadataClient *metadata.Client
+}
+
+func uploadOutputArtifacts(ctx context.Context, executorInput *pipelinespec.ExecutorInput, executorOutput *pipelinespec.ExecutorOutput, opts uploadOutputArtifactsOptions) ([]*metadata.OutputArtifact, error) {
+	// Register artifacts with MLMD.
+	outputArtifacts := make([]*metadata.OutputArtifact, 0, len(executorInput.GetOutputs().GetArtifacts()))
+	for name, artifactList := range executorInput.GetOutputs().GetArtifacts() {
+		if len(artifactList.Artifacts) == 0 {
+			continue
+		}
+		// TODO: Support multiple artifacts someday, probably through the v2 engine.
+		outputArtifact := artifactList.Artifacts[0]
+
+		// Merge executor output artifact info with executor input
+		if list, ok := executorOutput.Artifacts[name]; ok && len(list.Artifacts) > 0 {
+			mergeRuntimeArtifacts(list.Artifacts[0], outputArtifact)
+		}
+
+		// Upload artifacts from local path to remote storages.
+		localDir, err := localPathForURI(outputArtifact.Uri)
+		if err != nil {
+			glog.Warningf("Output Artifact %q does not have a recognized storage URI %q. Skipping uploading to remote storage.", name, outputArtifact.Uri)
+		} else {
+			blobKey, err := opts.bucketConfig.KeyFromURI(outputArtifact.Uri)
+			if err != nil {
+				return nil, fmt.Errorf("failed to upload output artifact %q: %w", name, err)
+			}
+			if err := objectstore.UploadBlob(ctx, opts.bucket, localDir, blobKey); err != nil {
+				//  We allow components to not produce output files
+				if errors.Is(err, os.ErrNotExist) {
+					glog.Warningf("Local filepath %q does not exist", localDir)
+				} else {
+					return nil, fmt.Errorf("failed to upload output artifact %q to remote storage URI %q: %w", name, outputArtifact.Uri, err)
+				}
+			}
+		}
+
+		// Write out the metadata.
+		metadataErr := func(err error) error {
+			return fmt.Errorf("unable to produce MLMD artifact for output %q: %w", name, err)
+		}
+		// TODO(neuromage): Consider batching these instead of recording one by one.
+		schema, err := getArtifactSchema(outputArtifact.GetType())
+		if err != nil {
+			return nil, fmt.Errorf("failed to determine schema for output %q: %w", name, err)
+		}
+		mlmdArtifact, err := opts.metadataClient.RecordArtifact(ctx, name, schema, outputArtifact, pb.Artifact_LIVE)
+		if err != nil {
+			return nil, metadataErr(err)
+		}
+		outputArtifacts = append(outputArtifacts, mlmdArtifact)
+	}
+	return outputArtifacts, nil
+}
+
+func downloadArtifacts(ctx context.Context, executorInput *pipelinespec.ExecutorInput, defaultBucket *blob.Bucket, defaultBucketConfig *objectstore.Config, namespace string, k8sClient *kubernetes.Clientset) error {
+	// Read input artifact metadata.
+	nonDefaultBuckets, err := fetchNonDefaultBuckets(ctx, executorInput.Inputs.Artifacts, defaultBucketConfig, namespace, k8sClient)
+	closeNonDefaultBuckets := func(buckets map[string]*blob.Bucket) {
+		for name, bucket := range nonDefaultBuckets {
+			if closeBucketErr := bucket.Close(); closeBucketErr != nil {
+				glog.Warningf("failed to close bucket %q: %q", name, err.Error())
+			}
+		}
+	}
+	defer closeNonDefaultBuckets(nonDefaultBuckets)
+	if err != nil {
+		return fmt.Errorf("failed to fetch non default buckets: %w", err)
+	}
+	for name, artifactList := range executorInput.Inputs.Artifacts {
+		// TODO(neuromage): Support concat-based placholders for arguments.
+		if len(artifactList.Artifacts) == 0 {
+			continue
+		}
+		inputArtifact := artifactList.Artifacts[0]
+		localPath, err := localPathForURI(inputArtifact.Uri)
+		if err != nil {
+			glog.Warningf("Input Artifact %q does not have a recognized storage URI %q. Skipping downloading to local path.", name, inputArtifact.Uri)
+			continue
+		}
+		// Copy artifact to local storage.
+		copyErr := func(err error) error {
+			return fmt.Errorf("failed to download input artifact %q from remote storage URI %q: %w", name, inputArtifact.Uri, err)
+		}
+		// TODO: Selectively copy artifacts for which .path was actually specified
+		// on the command line.
+		bucket := defaultBucket
+		bucketConfig := defaultBucketConfig
+		if !strings.HasPrefix(inputArtifact.Uri, defaultBucketConfig.PrefixedBucket()) {
+			nonDefaultBucketConfig, err := objectstore.ParseBucketConfigForArtifactURI(inputArtifact.Uri)
+			if err != nil {
+				return fmt.Errorf("failed to parse bucketConfig for output artifact %q with uri %q: %w", name, inputArtifact.GetUri(), err)
+			}
+			nonDefaultBucket, ok := nonDefaultBuckets[nonDefaultBucketConfig.PrefixedBucket()]
+			if !ok {
+				return fmt.Errorf("failed to get bucket when downloading input artifact %s with bucket key %s: %w", name, nonDefaultBucketConfig.PrefixedBucket(), err)
+			}
+			bucket = nonDefaultBucket
+			bucketConfig = nonDefaultBucketConfig
+		}
+		blobKey, err := bucketConfig.KeyFromURI(inputArtifact.Uri)
+		if err != nil {
+			return copyErr(err)
+		}
+		if err := objectstore.DownloadBlob(ctx, bucket, localPath, blobKey); err != nil {
+			return copyErr(err)
+		}
+
+	}
+	return nil
+}
+
+func fetchNonDefaultBuckets(ctx context.Context, artifacts map[string]*pipelinespec.ArtifactList, defaultBucketConfig *objectstore.Config, namespace string, k8sClient *kubernetes.Clientset) (buckets map[string]*blob.Bucket, err error) {
+	nonDefaultBuckets := make(map[string]*blob.Bucket)
+	for name, artifactList := range artifacts {
+		if len(artifactList.Artifacts) == 0 {
+			continue
+		}
+		// TODO: Support multiple artifacts someday, probably through the v2 engine.
+		artifact := artifactList.Artifacts[0]
+		if !strings.HasPrefix(artifact.Uri, defaultBucketConfig.PrefixedBucket()) {
+			nonDefaultBucketConfig, err := objectstore.ParseBucketConfigForArtifactURI(artifact.Uri)
+			if err != nil {
+				return nonDefaultBuckets, fmt.Errorf("failed to parse bucketConfig for output artifact %q with uri %q: %w", name, artifact.GetUri(), err)
+			}
+			nonDefaultBucket, err := objectstore.OpenBucket(ctx, k8sClient, namespace, nonDefaultBucketConfig)
+			if err != nil {
+				return nonDefaultBuckets, fmt.Errorf("failed to open bucket for output artifact %q with uri %q: %w", name, artifact.GetUri(), err)
+			}
+			nonDefaultBuckets[nonDefaultBucketConfig.PrefixedBucket()] = nonDefaultBucket
+		}
+
+	}
+	return nonDefaultBuckets, nil
+
+}
+
+// Add executor input placeholders to provided map.
+func getPlaceholders(executorInput *pipelinespec.ExecutorInput) (placeholders map[string]string, err error) {
+	defer func() {
+		if err != nil {
+			err = fmt.Errorf("failed to get placeholders: %w", err)
+		}
+	}()
+	placeholders = make(map[string]string)
+	executorInputJSON, err := protojson.Marshal(executorInput)
+	if err != nil {
+		return nil, fmt.Errorf("failed to convert ExecutorInput into JSON: %w", err)
+	}
+	placeholders["{{$}}"] = string(executorInputJSON)
+
+	// Read input artifact metadata.
+	for name, artifactList := range executorInput.Inputs.Artifacts {
+		if len(artifactList.Artifacts) == 0 {
+			continue
+		}
+		inputArtifact := artifactList.Artifacts[0]
+
+		// Prepare input uri placeholder.
+		key := fmt.Sprintf(`{{$.inputs.artifacts['%s'].uri}}`, name)
+		placeholders[key] = inputArtifact.Uri
+
+		localPath, err := localPathForURI(inputArtifact.Uri)
+		if err != nil {
+			// Input Artifact does not have a recognized storage URI
+			continue
+		}
+
+		// Prepare input path placeholder.
+		key = fmt.Sprintf(`{{$.inputs.artifacts['%s'].path}}`, name)
+		placeholders[key] = localPath
+	}
+
+	// Prepare output artifact placeholders.
+	for name, artifactList := range executorInput.GetOutputs().GetArtifacts() {
+		if len(artifactList.Artifacts) == 0 {
+			continue
+		}
+		outputArtifact := artifactList.Artifacts[0]
+		placeholders[fmt.Sprintf(`{{$.outputs.artifacts['%s'].uri}}`, name)] = outputArtifact.Uri
+
+		localPath, err := localPathForURI(outputArtifact.Uri)
+		if err != nil {
+			return nil, fmt.Errorf("resolve output artifact %q's local path: %w", name, err)
+		}
+		placeholders[fmt.Sprintf(`{{$.outputs.artifacts['%s'].path}}`, name)] = localPath
+	}
+
+	// Prepare input parameter placeholders.
+	for name, parameter := range executorInput.Inputs.ParameterValues {
+		key := fmt.Sprintf(`{{$.inputs.parameters['%s']}}`, name)
+		switch t := parameter.Kind.(type) {
+		case *structpb.Value_StringValue:
+			placeholders[key] = parameter.GetStringValue()
+		case *structpb.Value_NumberValue:
+			placeholders[key] = strconv.FormatFloat(parameter.GetNumberValue(), 'f', -1, 64)
+		case *structpb.Value_BoolValue:
+			placeholders[key] = strconv.FormatBool(parameter.GetBoolValue())
+		case *structpb.Value_ListValue:
+			b, err := json.Marshal(parameter.GetListValue())
+			if err != nil {
+				return nil, fmt.Errorf("failed to JSON-marshal list input parameter %q: %w", name, err)
+			}
+			placeholders[key] = string(b)
+		case *structpb.Value_StructValue:
+			b, err := json.Marshal(parameter.GetStructValue())
+			if err != nil {
+				return nil, fmt.Errorf("failed to JSON-marshal dict input parameter %q: %w", name, err)
+			}
+			placeholders[key] = string(b)
+		default:
+			return nil, fmt.Errorf("unknown PipelineSpec Value type %T", t)
+		}
+	}
+
+	// Prepare output parameter placeholders.
+	for name, parameter := range executorInput.GetOutputs().GetParameters() {
+		key := fmt.Sprintf(`{{$.outputs.parameters['%s'].output_file}}`, name)
+		placeholders[key] = parameter.OutputFile
+	}
+
+	return placeholders, nil
+}
+
+func getArtifactSchema(schema *pipelinespec.ArtifactTypeSchema) (string, error) {
+	switch t := schema.Kind.(type) {
+	case *pipelinespec.ArtifactTypeSchema_InstanceSchema:
+		return t.InstanceSchema, nil
+	case *pipelinespec.ArtifactTypeSchema_SchemaTitle:
+		return "title: " + t.SchemaTitle, nil
+	case *pipelinespec.ArtifactTypeSchema_SchemaUri:
+		return "", fmt.Errorf("SchemaUri is unsupported")
+	default:
+		return "", fmt.Errorf("unknown type %T in ArtifactTypeSchema %+v", t, schema)
+	}
+}
+
+func mergeRuntimeArtifacts(src, dst *pipelinespec.RuntimeArtifact) {
+	if len(src.Uri) > 0 {
+		dst.Uri = src.Uri
+	}
+
+	if src.Metadata != nil {
+		if dst.Metadata == nil {
+			dst.Metadata = src.Metadata
+		} else {
+			for k, v := range src.Metadata.Fields {
+				dst.Metadata.Fields[k] = v
+			}
+		}
+	}
+}
+
+func getExecutorOutputFile(path string) (*pipelinespec.ExecutorOutput, error) {
+	// collect user executor output file
+	executorOutput := &pipelinespec.ExecutorOutput{
+		ParameterValues: map[string]*structpb.Value{},
+		Artifacts:       map[string]*pipelinespec.ArtifactList{},
+	}
+
+	_, err := os.Stat(path)
+	if err != nil {
+		if os.IsNotExist(err) {
+			glog.Infof("output metadata file does not exist in %s", path)
+			// If file doesn't exist, return an empty ExecutorOutput.
+			return executorOutput, nil
+		} else {
+			return nil, fmt.Errorf("failed to stat output metadata file %q: %w", path, err)
+		}
+	}
+
+	b, err := ioutil.ReadFile(path)
+	if err != nil {
+		return nil, fmt.Errorf("failed to read output metadata file %q: %w", path, err)
+	}
+	glog.Infof("ExecutorOutput: %s", prettyPrint(string(b)))
+
+	if err := protojson.Unmarshal(b, executorOutput); err != nil {
+		return nil, fmt.Errorf("failed to unmarshall ExecutorOutput in file %q: %w", path, err)
+	}
+
+	return executorOutput, nil
+}
+
+func localPathForURI(uri string) (string, error) {
+	if strings.HasPrefix(uri, "gs://") {
+		return "/gcs/" + strings.TrimPrefix(uri, "gs://"), nil
+	}
+	if strings.HasPrefix(uri, "minio://") {
+		return "/minio/" + strings.TrimPrefix(uri, "minio://"), nil
+	}
+	if strings.HasPrefix(uri, "s3://") {
+		return "/s3/" + strings.TrimPrefix(uri, "s3://"), nil
+	}
+	return "", fmt.Errorf("failed to generate local path for URI %s: unsupported storage scheme", uri)
+}
+
+func prepareOutputFolders(executorInput *pipelinespec.ExecutorInput) error {
+	for name, parameter := range executorInput.GetOutputs().GetParameters() {
+		dir := filepath.Dir(parameter.OutputFile)
+		if err := os.MkdirAll(dir, 0644); err != nil {
+			return fmt.Errorf("failed to create directory %q for output parameter %q: %w", dir, name, err)
+		}
+	}
+
+	for name, artifactList := range executorInput.GetOutputs().GetArtifacts() {
+		if len(artifactList.Artifacts) == 0 {
+			continue
+		}
+		outputArtifact := artifactList.Artifacts[0]
+
+		localPath, err := localPathForURI(outputArtifact.Uri)
+		if err != nil {
+			return fmt.Errorf("failed to generate local storage path for output artifact %q: %w", name, err)
+		}
+
+		if err := os.MkdirAll(filepath.Dir(localPath), 0644); err != nil {
+			return fmt.Errorf("unable to create directory %q for output artifact %q: %w", filepath.Dir(localPath), name, err)
+		}
+	}
+
+	return nil
 }
